I was trying to build an XML file on the fly from a JSON sort of definition give in the script itself.
That's turning out to be a silly idea in retrospect. (1) It's taking too much time to code this, and (2) more importantly there's no use to this. Much easier to feed the XML directly as that's what Azure accepts anyway. Why am I introducing an artificial syntax in between and possibly errors from my side when translating? I realized then when I had to code the <Gateway> and <ConnectionsToLocalNetwork> parts and it struck me that I would to make corresponding definitions for that in the JSON definition. No point!

Rather than waste the work I am creating a trash1 branch and shoving it off there.

The way I create XML is quite good. I had to learn it for this post, so I don't want to erase it either. :)